requirements.txt
fastapi==0.115.11
mcp==1.5.0
python-dotenv==1.0.1
ninja2==0.1
uvicorn==0.34.0
